v0.5.3
* Fixed issues with v0.11 stream implementation #73
* Fixed issues with pause/resume and data events in v0.10 #69
* Fixed the double invoking of done callback when parsing files #68
* Refactored tests
