+++ .travis.yml
@@ -1,5 +1,6 @@
 language: node_js
 node_js:
   - "0.10"
+  - "0.11"
 before_script:
   - npm install -g grunt-cli
