+++ benchmark/benchmark.js
@@ -1,9 +1,7 @@
 var fastCsv = require("../lib"),
     csv = require("csv"),
     path = require("path"),
-    fs = require("fs"),
-    COUNT = 10000000,
-    TEST_FILE = path.resolve(__dirname, "./assets/" + COUNT + ".csv");
+    fs = require("fs");
 
 
 function camelize(str) {
@@ -12,10 +10,11 @@ function camelize(str) {
     });
 }
 
-function benchmarkFastCsv(done) {
-    var count = 0;
+function benchmarkFastCsv(num, done) {
+    var count = 0,
+        file = path.resolve(__dirname, "./assets/" + num + ".csv");
     fastCsv
-        .fromPath(TEST_FILE, {headers: true})
+        .fromPath(file, {headers: true})
         .transform(function (data) {
             var ret = {};
             ["first_name", "last_name", "email_address"].forEach(function (prop) {
@@ -28,8 +27,8 @@ function benchmarkFastCsv(done) {
             count++;
         })
         .on("end", function (count) {
-            if (count !== COUNT) {
-                done(new Error("Error expected " + COUNT + " got " + count));
+            if (count !== num) {
+                done(new Error("Error expected " + num + " got " + count));
             } else {
                 done();
             }
@@ -40,9 +39,10 @@ function benchmarkFastCsv(done) {
 
 }
 
-function benchmarkCsv(done) {
-    var count = 0;
-    fs.createReadStream(TEST_FILE)
+function benchmarkCsv(num, done) {
+    var count = 0,
+        file = path.resolve(__dirname, "./assets/" + num + ".csv");
+    fs.createReadStream(file)
         .pipe(csv.parse({columns: true}))
         .pipe(csv.transform(function (data) {
             var ret = {};
@@ -56,8 +56,8 @@ function benchmarkCsv(done) {
             count++;
         })
         .on('end', function () {
-            if (count !== COUNT) {
-                done(new Error("Error expected " + COUNT + " got " + count));
+            if (count !== num) {
+                done(new Error("Error expected " + num + " got " + count));
             } else {
                 done();
             }
@@ -67,26 +67,26 @@ function benchmarkCsv(done) {
         });
 }
 
-function benchmark(title, m, done) {
+function benchmark(title, num, m, done) {
     var start = new Date(), runStart = start;
-    m(function (err) {
+    m(num, function (err) {
         if (err) {
             done(err);
         } else {
-            console.log("%s: RUN 1 %dms", title, (new Date() - runStart));
+            console.log("%s: RUN(%d lines) 1 %dms", title, num, (new Date() - runStart));
             runStart = new Date();
-            m(function (err) {
+            m(num, function (err) {
                 if (err) {
                     done(err);
                 } else {
-                    console.log("%s: RUN 2 %dms", title, (new Date() - runStart));
+                    console.log("%s: RUN(%d lines) 2 %dms", title, num, (new Date() - runStart));
                     runStart = new Date();
-                    m(function (err) {
+                    m(num, function (err) {
                         if (err) {
                             done(err);
                         } else {
-                            console.log("%s: RUN 3 %dms", title, (new Date() - runStart));
-                            console.log("%s: 3xAVG %dms", title, (new Date() - start) / 3);
+                            console.log("%s: RUN(%d lines) 3 %dms", title, num, (new Date() - runStart));
+                            console.log("%s: 3xAVG for %d lines %dms", title, num, (new Date() - start) / 3);
                             done();
                         }
 
@@ -97,14 +97,43 @@ function benchmark(title, m, done) {
     });
 }
 
-benchmark('fast-csv', benchmarkFastCsv, function (err) {
-    if (err) {
-        console.log(err.stack);
-    } else {
-        benchmark('csv', benchmarkCsv, function (err) {
-            if (err) {
-                console.log(err.stack);
-            }
-        });
-    }
-});
+function runBenchmarks(num, cb) {
+    console.log("RUNNING %d.csv benchmarks", num);
+    benchmark('fast-csv', num, benchmarkFastCsv, function (err) {
+        if (err) {
+            cb(err);
+        } else {
+            benchmark('csv', num, benchmarkCsv, function (err) {
+                if (err) {
+                    cb(err);
+                } else {
+                    console.log("");
+                    cb();
+                }
+            });
+        }
+    });
+}
+
+module.exports = function benchmarks(cb) {
+    runBenchmarks(20000, function (err) {
+        if (err) {
+            cb(err);
+        } else {
+            runBenchmarks(50000, function (err) {
+                if (err) {
+                    cb(err);
+                } else {
+                    runBenchmarks(100000, function (err) {
+                        if (err) {
+                            cb(err);
+                        } else {
+                            cb(null);
+                        }
+                    });
+                }
+            });
+        }
+    });
+};
+
