+++ benchmark/createData.js
@@ -20,6 +20,7 @@ while (++i < l) {
     lines.push(lineOptions[i % lineOptionsLength]);
 }
 
+console.log("writing %d lines to  %d.csv", l, l);
 fastCsv
     .writeToPath(path.resolve(__dirname, "./assets/20000.csv"), lines)
     .on('finish', function () {
@@ -28,6 +29,7 @@ fastCsv
             lines.push(lineOptions[i % lineOptionsLength]);
         }
 
+        console.log("writing %d lines to  %d.csv", l - 1, l - 1);
         fastCsv
             .writeToPath(path.resolve(__dirname, "./assets/50000.csv"), lines)
             .on('finish', function () {
@@ -35,7 +37,7 @@ fastCsv
                 while (++i < l) {
                     lines.push(lineOptions[i % lineOptionsLength]);
                 }
-
+                console.log("writing %d lines to  %d.csv", l - 2, l - 2);
                 fastCsv
                     .writeToPath(path.resolve(__dirname, "./assets/100000.csv"), lines)
                     .on('finish', function () {
@@ -43,6 +45,7 @@ fastCsv
                         while (++i < l) {
                             lines.push(lineOptions[i % lineOptionsLength]);
                         }
+                        console.log("writing %d lines to  %d.csv", l - 3, l - 3);
                         fastCsv.writeToPath(path.resolve(__dirname, "./assets/1000000.csv"), lines)
                             .on("finish", function () {
                                 process.exit();
