+++ lib/parser/parser_stream.js
@@ -41,24 +41,8 @@ function ParserStream(options) {
 util.inherits(ParserStream, stream.Transform);
 
 var origOn = ParserStream.prototype.on,
-    origPause = ParserStream.prototype.pause,
-    origResume = ParserStream.prototype.resume,
     origEmit = ParserStream.prototype.emit;
 
-function pause() {
-    spreadArgs(origPause, arguments, this);
-    this.paused = true;
-    this.pause = pause;
-}
-
-function resume() {
-    spreadArgs(origResume, arguments, this);
-    this.paused = false;
-    if (this.__pausedDone) {
-        this.__pausedDone();
-    }
-    this.resume = resume;
-}
 
 extended(ParserStream).extend({
 
@@ -104,7 +88,7 @@ extended(ParserStream).extend({
                         cb(err);
                     } else {
                         if (dataRow) {
-                            if (!self.paused) {
+                            if (!self.isPaused()) {
                                 self.__emitRecord(dataRow, count);
                             } else {
                                 self.__buffered.push([dataRow, count]);
@@ -206,7 +190,7 @@ extended(ParserStream).extend({
                     done(err);
                 } else {
                     self.lines = lineData;
-                    if (!self.paused) {
+                    if (!self.isPaused()) {
                         done();
                     } else {
                         self.__pausedDone = done;
@@ -215,7 +199,7 @@ extended(ParserStream).extend({
             });
         } else {
             this.lines = lineData;
-            if (!this.paused) {
+            if (!this.isPaused()) {
                 done();
             } else {
                 this.__pausedDone = done;
@@ -225,9 +209,13 @@ extended(ParserStream).extend({
     },
 
     __doFlush: function (callback) {
-        //increment row count so we aren't 0 based
-        this.emit("end");
-        callback();
+        try {
+            //increment row count so we aren't 0 based
+            this.emit("end");
+            callback();
+        } catch (e) {
+            callback(e);
+        }
     },
 
     _flush: function (callback) {
@@ -236,7 +224,7 @@ extended(ParserStream).extend({
             this._parse(this.lines, false, function (err) {
                 if (err) {
                     callback(err);
-                } else if (!self.paused) {
+                } else if (!self.isPaused()) {
                     self.__doFlush(callback);
                 } else {
                     self.__pausedDone = function () {
@@ -245,7 +233,7 @@ extended(ParserStream).extend({
                 }
             });
         } else {
-            if (!this.paused) {
+            if (!this.isPaused()) {
                 this.__doFlush(callback);
             } else {
                 this.__pausedDone = function () {
@@ -263,11 +251,29 @@ extended(ParserStream).extend({
         return next(null, data);
     },
 
-    pause: function () {
-        if (!this.paused) {
-            this.paused = true;
-            this.emit("pause");
+    __flushPausedBuffer: function () {
+        var buffered = this.__buffered, l = buffered.length;
+        if (l) {
+            var entry;
+            while (buffered.length) {
+                entry = buffered.shift();
+                this.__emitRecord(entry[0], entry[1]);
+                //handle case where paused is called while emitting data
+                if (this.isPaused()) {
+                    return;
+                }
+            }
+            buffered.length = 0;
         }
+        if (this.__pausedDone) {
+            var done = this.__pausedDone;
+            this.__pausedDone = null;
+            done();
+        }
+    },
+
+    isPaused: function () {
+        return !!this.__paused;
     },
 
     emit: function (event) {
@@ -277,32 +283,13 @@ extended(ParserStream).extend({
                 spreadArgs(origEmit, ["end", ++this._rowCount], this);
             }
         } else {
+            if (event === "pause") {
+                this.__paused = true;
+            } else if (event === "resume") {
+                this.__paused = false;
+                this.__flushPausedBuffer();
+            }
             spreadArgs(origEmit, arguments, this);
-        }
-    },
-
-    resume: function () {
-        if (this.paused) {
-            this.paused = false;
-            var buffered = this.__buffered, l = buffered.length;
-            if (l) {
-                var entry;
-                while (buffered.length) {
-                    entry = buffered.shift();
-                    this.__emitRecord(entry[0], entry[1]);
-                    //handle case where paused is called while emitting data
-                    if (this.paused) {
-                        return;
-                    }
-                }
-                buffered.length = 0;
-            }
-            if (this.__pausedDone) {
-                var done = this.__pausedDone;
-                this.__pausedDone = null;
-                done();
-            }
-            this.emit("resume");
         }
     },
