+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "fast-csv",
-    "version": "0.5.2",
+    "version": "0.5.3",
     "description": "CSV parser and writer",
     "main": "index.js",
     "scripts": {
